# Speed up S3 Client tests with mocks