The tests creating an `s3_client` were taking ~10 seconds to run 3 tests. They were waiting on a socket connection to a fake region to fail.

This change:
- makes the AWS utils tests more granular
- adds `moto.mock_aws` around the slow client creation.

On my machine it brings the runtime of `test_utils_aws.py` down from 9.6s to 1.3s.

And the runtime for the entire `tests/` directory down from 18.5s to 8.4s.

### Before
![image](https://github.com/user-attachments/assets/0e6db5ad-4f5d-4825-8b7a-c0f01db740a8)


### After
![image](https://github.com/user-attachments/assets/189790bb-04e6-4df1-8dd8-e78bc4cef95e)



<!--
See https://github.com/blog/2111-issue-and-pull-request-templates for more
information on pull request templates.

-->


<!-- readthedocs-preview opendatacube start -->
----
📚 Documentation preview 📚: https://opendatacube--1878.org.readthedocs.build/en/1878/

<!-- readthedocs-preview opendatacube end -->